updated SpMV cached example
