+++ examples/spmv_compute.c
@@ -1,9 +1,27 @@
 // Generated by the Tensor Algebra Compiler (tensor-compiler.org)
-for (int32_t iA = 0; iA < A1_size; iA++) {
-  double tj = 0;
-  for (int32_t pA2 = A2_pos[iA]; pA2 < A2_pos[iA + 1]; pA2++) {
-    int32_t jA = A2_idx[pA2];
-    tj += A_vals[pA2] * x_vals[jA];
+// taco "y(i)=A(i,j)*x(j)" -f=y:d:0 -f=A:ds:0,1 -f=x:d:0 -write-source=taco_kernel.c -write-compute=taco_compute.c -write-assembly=taco_assembly.c
+
+int compute(taco_tensor_t *y, taco_tensor_t *A, taco_tensor_t *x) {
+  int y1_dimension = (int)(y->dimensions[0]);
+  double* restrict y_vals = (double*)(y->vals);
+  int A1_dimension = (int)(A->dimensions[0]);
+  int* restrict A2_pos = (int*)(A->indices[1][0]);
+  int* restrict A2_crd = (int*)(A->indices[1][1]);
+  double* restrict A_vals = (double*)(A->vals);
+  int x1_dimension = (int)(x->dimensions[0]);
+  double* restrict x_vals = (double*)(x->vals);
+
+  #pragma omp parallel for schedule(static)
+  for (int32_t py = 0; py < y1_dimension; py++) {
+    y_vals[py] = 0.0;
   }
-  y_vals[iA] = tj;
+
+  #pragma omp parallel for schedule(runtime)
+  for (int32_t i = 0; i < A1_dimension; i++) {
+    for (int32_t jA = A2_pos[i]; jA < A2_pos[(i + 1)]; jA++) {
+      int32_t j = A2_crd[jA];
+      y_vals[i] = y_vals[i] + A_vals[jA] * x_vals[j];
+    }
+  }
+  return 0;
 }
